refactor: move wallet/connection util functions into `devtools-solana` package for reusability (#1504)
