+++ examples/lzapp-migration/tasks/common/utils.ts
@@ -1,34 +1,13 @@
-import assert from 'assert'
+import { PublicKey } from '@solana/web3.js'
 
-import { Keypair, PublicKey } from '@solana/web3.js'
-
-import {
-    OmniPoint,
-    OmniSigner,
-    OmniTransactionReceipt,
-    OmniTransactionResponse,
-    firstFactory,
-    formatEid,
-} from '@layerzerolabs/devtools'
+import { OmniPoint, firstFactory } from '@layerzerolabs/devtools'
 import { createConnectedContractFactory } from '@layerzerolabs/devtools-evm-hardhat'
-import {
-    OmniSignerSolana,
-    OmniSignerSolanaSquads,
-    createConnectionFactory,
-    createRpcUrlFactory,
-} from '@layerzerolabs/devtools-solana'
-import { ChainType, EndpointId, endpointIdToChainType } from '@layerzerolabs/lz-definitions'
+import { createSolanaConnectionFactory, createSolanaSignerFactory } from '@layerzerolabs/devtools-solana'
 import { IOApp } from '@layerzerolabs/ua-devtools'
 import { createOAppFactory } from '@layerzerolabs/ua-devtools-evm'
 import { createOFTFactory } from '@layerzerolabs/ua-devtools-solana'
 
-export const createSolanaConnectionFactory = () =>
-    createConnectionFactory(
-        createRpcUrlFactory({
-            [EndpointId.SOLANA_V2_MAINNET]: process.env.RPC_URL_SOLANA,
-            [EndpointId.SOLANA_V2_TESTNET]: process.env.RPC_URL_SOLANA_TESTNET,
-        })
-    )
+export { createSolanaConnectionFactory }
 
 export const createSdkFactory = (
     userAccount: PublicKey,
@@ -69,19 +48,4 @@ export const createSdkFactory = (
     return firstFactory<[OmniPoint], IOApp>(evmSdkfactory, solanaSdkFactory)
 }
 
-export const createSolanaSignerFactory = (
-    wallet: Keypair,
-    connectionFactory = createSolanaConnectionFactory(),
-    multisigKey?: PublicKey
-) => {
-    return async (eid: EndpointId): Promise<OmniSigner<OmniTransactionResponse<OmniTransactionReceipt>>> => {
-        assert(
-            endpointIdToChainType(eid) === ChainType.SOLANA,
-            `Solana signer factory can only create signers for Solana networks. Received ${formatEid(eid)}`
-        )
-
-        return multisigKey
-            ? new OmniSignerSolanaSquads(eid, await connectionFactory(eid), multisigKey, wallet)
-            : new OmniSignerSolana(eid, await connectionFactory(eid), wallet)
-    }
-}
+export { createSolanaSignerFactory }
