+++ examples/oft-solana/tasks/common/utils.ts
@@ -1,30 +1,17 @@
-import assert from 'assert'
+import { Connection, PublicKey } from '@solana/web3.js'
 
-import { Connection, Keypair, PublicKey } from '@solana/web3.js'
-
-import {
-    OmniPoint,
-    OmniSigner,
-    OmniTransactionReceipt,
-    OmniTransactionResponse,
-    firstFactory,
-    formatEid,
-} from '@layerzerolabs/devtools'
+import { OmniPoint, firstFactory } from '@layerzerolabs/devtools'
 import { createConnectedContractFactory } from '@layerzerolabs/devtools-evm-hardhat'
-import {
-    OmniSignerSolana,
-    OmniSignerSolanaSquads,
-    createConnectionFactory,
-    createRpcUrlFactory,
-} from '@layerzerolabs/devtools-solana'
+import { createSolanaConnectionFactory, createSolanaSignerFactory } from '@layerzerolabs/devtools-solana'
 import { createLogger } from '@layerzerolabs/io-devtools'
-import { ChainType, EndpointId, endpointIdToChainType, endpointIdToNetwork } from '@layerzerolabs/lz-definitions'
+import { EndpointId, endpointIdToNetwork } from '@layerzerolabs/lz-definitions'
 import { UlnProgram } from '@layerzerolabs/lz-solana-sdk-v2'
 import { Options } from '@layerzerolabs/lz-v2-utilities'
 import { IOApp } from '@layerzerolabs/ua-devtools'
 import { createOAppFactory } from '@layerzerolabs/ua-devtools-evm'
 import { createOFTFactory } from '@layerzerolabs/ua-devtools-solana'
 
+export { createSolanaConnectionFactory }
 const logger = createLogger()
 
 export const deploymentMetadataUrl = 'https://metadata.layerzero-api.com/v1/metadata/deployments'
@@ -46,14 +33,6 @@ export async function getBlockExplorerLink(srcEid: number, txHash: string): Prom
     }
     return
 }
-
-export const createSolanaConnectionFactory = () =>
-    createConnectionFactory(
-        createRpcUrlFactory({
-            [EndpointId.SOLANA_V2_MAINNET]: process.env.RPC_URL_SOLANA,
-            [EndpointId.SOLANA_V2_TESTNET]: process.env.RPC_URL_SOLANA_TESTNET,
-        })
-    )
 
 export const createSdkFactory = (
     userAccount: PublicKey,
@@ -94,22 +73,7 @@ export const createSdkFactory = (
     return firstFactory<[OmniPoint], IOApp>(evmSdkfactory, solanaSdkFactory)
 }
 
-export const createSolanaSignerFactory = (
-    wallet: Keypair,
-    connectionFactory = createSolanaConnectionFactory(),
-    multisigKey?: PublicKey
-) => {
-    return async (eid: EndpointId): Promise<OmniSigner<OmniTransactionResponse<OmniTransactionReceipt>>> => {
-        assert(
-            endpointIdToChainType(eid) === ChainType.SOLANA,
-            `Solana signer factory can only create signers for Solana networks. Received ${formatEid(eid)}`
-        )
-
-        return multisigKey
-            ? new OmniSignerSolanaSquads(eid, await connectionFactory(eid), multisigKey, wallet)
-            : new OmniSignerSolana(eid, await connectionFactory(eid), wallet)
-    }
-}
+export { createSolanaSignerFactory }
 
 export function uint8ArrayToHex(uint8Array: Uint8Array, prefix = false): string {
     const hexString = Buffer.from(uint8Array).toString('hex')
