+++ package.json
@@ -34,6 +34,7 @@
     ]
   },
   "resolutions": {
+    "@solana/web3.js": "^1.95.8",
     "@ton/ton": "npm:@layerzerolabs/ton@15.2.0-rc.3",
     "es5-ext": "git://github.com/LayerZero-Labs/es5-ext",
     "ethers": "^5.7.2",
@@ -68,5 +69,13 @@
   "packageManager": "pnpm@8.15.6",
   "engines": {
     "node": ">=18.16.0"
+  },
+  "pnpm": {
+    "overrides": {
+      "@solana/web3.js": "^1.95.8"
+    }
+  },
+  "overrides": {
+    "@solana/web3.js": "^1.95.8"
   }
 }
