+++ packages/devtools-solana/src/index.ts
@@ -2,3 +2,4 @@ export * from './common'
 export * from './connection'
 export * from './omnigraph'
 export * from './transactions'
+export * from './wallet'
